add some messages, split can and compute, remove some header circ
