+++ Core/Src/adi_interaction.c
@@ -307,6 +307,11 @@ void read_status_registers(cell_asic chips[NUM_CHIPS])
 	read_adbms_data(chips, RDSTATE, Status, E);
 }
 
+void read_status_register_c(cell_asic chips[NUM_CHIPS])
+{
+	read_adbms_data(chips, RDSTATC, Status, C);
+}
+
 void read_status_aux_registers(cell_asic chips[NUM_CHIPS])
 {
 	read_adbms_data(chips, RDASALL, Rdasall, ALL_GRP);
