+++ Core/Src/stateMachine.c
@@ -1,6 +1,8 @@
 #include "stateMachine.h"
-#include <stdio.h>
-#include <stdlib.h>
+
+#include "can_messages.h"
+#include "compute.h"
+#include "segment.h"
 
 #define MIN(a, b) (((a) < (b)) ? (a) : (b))
 extern UART_HandleTypeDef huart4;
@@ -122,7 +124,7 @@ void handle_charging(acc_data_t *bmsdata)
 			compute_enable_charging(true);
 		else {
 			compute_enable_charging(false);
-			compute_send_charging_message(0, 0, bmsdata);
+			send_charging_message(0, 0, bmsdata);
 		}
 
 		/* Check if we should balance */
@@ -134,7 +136,7 @@ void handle_charging(acc_data_t *bmsdata)
 		/* Send CAN message, but not too often */
 		if (is_timer_expired(&charger_message_timer) ||
 		    !is_timer_active(&charger_message_timer)) {
-			compute_send_charging_message(
+			send_charging_message(
 				(MAX_CHARGE_VOLT *
 				 (NUM_CELLS_ALPHA + NUM_CELLS_BETA) *
 				 NUM_CHIPS),
@@ -341,15 +343,13 @@ bool sm_fault_eval(fault_eval_t *item)
 		if (!fault_present) {
 			printf("\t\t\t*******Fault cleared: %s\r\n", item->id);
 			cancel_timer(&item->timer);
-			compute_send_fault_timer_message(0, item->code,
-							 item->data_1);
+			send_fault_timer_message(0, item->code, item->data_1);
 			return 0;
 		}
 
 		if (is_timer_expired(&item->timer) && fault_present) {
 			printf("\t\t\t*******Faulted: %s\r\n", item->id);
-			compute_send_fault_timer_message(2, item->code,
-							 item->data_1);
+			send_fault_timer_message(2, item->code, item->data_1);
 			return item->code;
 		}
 
@@ -361,7 +361,7 @@ bool sm_fault_eval(fault_eval_t *item)
 	else if (!is_timer_active(&item->timer) && fault_present) {
 		printf("\t\t\t*******Starting fault timer: %s\r\n", item->id);
 		start_timer(&item->timer, item->timeout);
-		compute_send_fault_timer_message(1, item->code, item->data_1);
+		send_fault_timer_message(1, item->code, item->data_1);
 
 		return 0;
 	}
